Split SpatialFunctionsAcceptanceTest into multiple files.
