+++ enterprise/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/SpatialDistanceAcceptanceTest.scala
@@ -1,0 +1,293 @@
+/*
+ * Copyright (c) 2002-2018 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.internal.cypher.acceptance
+
+import org.neo4j.cypher.ExecutionEngineFunSuite
+import org.neo4j.internal.cypher.acceptance.CypherComparisonSupport._
+import org.neo4j.values.storable.{CoordinateReferenceSystem, Values}
+
+class SpatialDistanceAcceptanceTest extends ExecutionEngineFunSuite with CypherComparisonSupport {
+
+  val pointConfig = Configs.Interpreted - Configs.Version2_3
+  val distanceConfig = Configs.Interpreted - Configs.OldAndRule
+
+  test("distance function should work on co-located points") {
+    val result = executeWith(pointConfig, "WITH point({latitude: 12.78, longitude: 56.7}) as point RETURN distance(point,point) as dist",
+      planComparisonStrategy = ComparePlansWithAssertion(_ should useOperatorWithText("Projection", "point", "dist"),
+        expectPlansToFail = Configs.AllRulePlanners))
+
+    result.toList should equal(List(Map("dist" -> 0.0)))
+  }
+
+  test("distance function should work on nearby cartesian points") {
+    val result = executeWith(pointConfig,
+      """
+        |WITH point({x: 2.3, y: 4.5, crs: 'cartesian'}) as p1, point({x: 1.1, y: 5.4, crs: 'cartesian'}) as p2
+        |RETURN distance(p1,p2) as dist
+      """.stripMargin,
+      planComparisonStrategy = ComparePlansWithAssertion(_ should useOperatorWithText("Projection", "p1", "p2", "dist"),
+        expectPlansToFail = Configs.AllRulePlanners))
+
+    result.columnAs("dist").next().asInstanceOf[Double] should equal(1.5)
+  }
+
+  test("distance function should work on nearby points") {
+    val result = executeWith(pointConfig,
+      """
+        |WITH point({longitude: 12.78, latitude: 56.7}) as p1, point({latitude: 56.71, longitude: 12.79}) as p2
+        |RETURN distance(p1,p2) as dist
+      """.stripMargin,
+      planComparisonStrategy = ComparePlansWithAssertion(_ should useOperatorWithText("Projection", "p1", "p2", "dist"),
+        expectPlansToFail = Configs.AllRulePlanners))
+
+    Math.round(result.columnAs("dist").next().asInstanceOf[Double]) should equal(1270)
+  }
+
+  test("distance function should work on distant points") {
+    val result = executeWith(pointConfig,
+      """
+        |WITH point({latitude: 56.7, longitude: 12.78}) as p1, point({longitude: -51.9, latitude: -16.7}) as p2
+        |RETURN distance(p1,p2) as dist
+      """.stripMargin,
+    planComparisonStrategy = ComparePlansWithAssertion(_ should useOperatorWithText("Projection", "p1", "p2", "dist"),
+      expectPlansToFail = Configs.AllRulePlanners))
+
+    Math.round(result.columnAs("dist").next().asInstanceOf[Double]) should equal(10116214)
+  }
+
+  test("distance function should not fail if provided with points from different CRS") {
+    val localConfig = pointConfig - Configs.OldAndRule
+      val res = executeWith(localConfig,
+        """WITH point({x: 2.3, y: 4.5, crs: 'cartesian'}) as p1, point({longitude: 1.1, latitude: 5.4, crs: 'WGS-84'}) as p2
+        |RETURN distance(p1,p2) as dist""".stripMargin)
+    res.columnAs[AnyRef]("dist").next() should be (null)
+  }
+
+  test("distance function should measure distance from Copenhagen train station to Neo4j in Malmö") {
+    val result = executeWith(pointConfig,
+      """
+        |WITH point({latitude: 55.672874, longitude: 12.564590}) as p1, point({latitude: 55.611784, longitude: 12.994341}) as p2
+        |RETURN distance(p1,p2) as dist
+      """.stripMargin,
+      planComparisonStrategy = ComparePlansWithAssertion(_ should useOperatorWithText("Projection", "p1", "p2","dist"),
+        expectPlansToFail = Configs.AllRulePlanners))
+
+    Math.round(result.columnAs("dist").next().asInstanceOf[Double]) should equal(27842)
+  }
+
+  test("distance function should work with two null inputs") {
+    val result = executeWith(pointConfig, "RETURN distance(null, null) as dist")
+    result.toList should equal(List(Map("dist" -> null)))
+  }
+
+  test("distance function should return null with lhs null input") {
+    val result = executeWith(pointConfig,
+      """
+        |WITH point({latitude: 55.672874, longitude: 12.564590}) as p1
+        |RETURN distance(null, p1) as dist
+      """.stripMargin)
+    result.toList should equal(List(Map("dist" -> null)))
+  }
+
+  test("distance function should return null with rhs null input") {
+    val result = executeWith(pointConfig,
+      """
+        |WITH point({latitude: 55.672874, longitude: 12.564590}) as p1
+        |RETURN distance(p1, null) as dist
+      """.stripMargin)
+    result.toList should equal(List(Map("dist" -> null)))
+  }
+
+  test("distance function should return null if a point is null") {
+    var result = executeWith(pointConfig,
+      "RETURN distance(point({latitude:3,longitude:7}),point({latitude:null, longitude:3})) as dist;")
+    result.toList should equal(List(Map("dist" -> null)))
+
+    result = executeWith(pointConfig,
+      "RETURN distance(point({latitude:3,longitude:null}),point({latitude:7, longitude:3})) as dist;")
+    result.toList should equal(List(Map("dist" -> null)))
+
+    result = executeWith(pointConfig,
+      "RETURN distance(point({x:3,y:7}),point({x:null, y:3})) as dist;")
+    result.toList should equal(List(Map("dist" -> null)))
+
+    result = executeWith(pointConfig,
+      "RETURN distance(point({x:3,y:null}),point({x:7, y:3})) as dist;")
+    result.toList should equal(List(Map("dist" -> null)))
+  }
+
+  test("distance function should fail on wrong type") {
+    val config = Configs.AbsolutelyAll + TestConfiguration(Versions.Default, Planners.Default, Runtimes.Default) - Configs.Version2_3
+    failWithError(config, "RETURN distance(1, 2) as dist", List("Type mismatch: expected Point or Geometry but was Integer"))
+  }
+
+  test("points with distance query and mixed crs") {
+    // Given
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 55.7, x: 11.78, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({latitude: 56.7, longitude: 12.78, crs: 'WGS-84'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({latitude: 55.7, longitude: 11.78, crs: 'WGS-84'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({latitude: 50.7, longitude: 12.78, crs: 'WGS-84'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({latitude: 56.7, longitude: 10.78, crs: 'WGS-84'})")
+
+    val query =
+      """CYPHER MATCH (p:Place)
+        |WHERE distance(p.location, point({latitude: 55.7, longitude: 11.78, crs: 'WGS-84'})) < 1000
+        |RETURN p.location as point
+      """.stripMargin
+    // When
+    val result = executeWith(distanceConfig, query)
+
+    // Then
+    val plan = result.executionPlanDescription()
+    plan should useOperatorWithText("Projection", "point")
+    plan should useOperatorWithText("Filter", "distance")
+    result.toList should equal(List(Map("point" -> Values.pointValue(CoordinateReferenceSystem.WGS84, 11.78, 55.7))))
+  }
+
+  test("indexed points with distance query and points within bbox") {
+    // Given
+    graph.createIndex("Place", "location")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: -10, x: -10, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: -10, x: 10, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 10, x: -10, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 10, x: 10, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: -10, x: 0, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 10, x: 0, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 0, x: -10, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 0, x: 10, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 0, x: 0, crs: 'cartesian'})")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 9.99, x: 0, crs: 'cartesian'})")
+    Range(11, 100).foreach(i => graph.execute(s"CREATE (p:Place) SET p.location = point({y: $i, x: $i, crs: 'cartesian'})"))
+
+    // <=
+    {
+      val query =
+        s"""CYPHER MATCH (p:Place)
+           |WHERE distance(p.location, point({x: 0, y: 0, crs: 'cartesian'})) <= 10
+           |RETURN p.location as point
+        """.stripMargin
+      // When
+      val result = executeWith(distanceConfig, query)
+
+      // Then
+      val plan = result.executionPlanDescription()
+      plan should useOperatorWithText("Projection", "point")
+      plan should useOperatorWithText("Filter", "distance")
+      plan should useOperatorWithText("NodeIndexSeekByRange", ":Place(location)", "distance", "<= ")
+      result.toList.toSet should equal(Set(
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 10, 0)),
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 0, 10)),
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, -10, 0)),
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 0, -10)),
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 0, 0)),
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 0, 9.99))
+      ))
+    }
+    // <
+    {
+      val query =
+        s"""CYPHER MATCH (p:Place)
+           |WHERE distance(p.location, point({x: 0, y: 0, crs: 'cartesian'})) < 10
+           |RETURN p.location as point
+        """.stripMargin
+      // When
+      val result = executeWith(distanceConfig, query)
+
+      // Then
+      val plan = result.executionPlanDescription()
+      plan should useOperatorWithText("Projection", "point")
+      plan should useOperatorWithText("Filter", "distance")
+      plan should useOperatorWithText("NodeIndexSeekByRange", ":Place(location)", "distance", "< ")
+      result.toList.toSet should equal(Set(
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 0, 0)),
+        Map("point" -> Values.pointValue(CoordinateReferenceSystem.Cartesian, 0, 9.99))
+      ))
+    }
+  }
+
+  test("invalid location with index") {
+    // Given
+    graph.createIndex("Place", "location")
+    graph.execute("CREATE (p:Place) SET p.location = 5")
+    Range(11, 100).foreach(i => graph.execute(s"CREATE (p:Place) SET p.location = point({y: $i, x: $i, crs: 'cartesian'})"))
+
+    val query =
+      s"""CYPHER MATCH (p:Place)
+         |WHERE distance(p.location, point({x: 0, y: 0, crs: 'cartesian'})) <= 10
+         |RETURN p.location as point
+        """.stripMargin
+    // When
+    val result = executeWith(distanceConfig, query)
+
+    // Then
+    val plan = result.executionPlanDescription()
+    plan should useOperatorWithText("Projection", "point")
+    plan should useOperatorWithText("Filter", "distance")
+    plan should useOperatorWithText("NodeIndexSeekByRange", ":Place(location)", "distance", "<= ")
+    result.toList.toSet should equal(Set.empty)
+  }
+
+  test("invalid location without index") {
+    // Given
+    graph.execute("CREATE (p:Place) SET p.location = 5")
+    Range(11, 100).foreach(i => graph.execute(s"CREATE (p:Place) SET p.location = point({y: $i, x: $i, crs: 'cartesian'})"))
+
+    val query =
+      s"""CYPHER MATCH (p:Place)
+         |WHERE distance(p.location, point({x: 0, y: 0, crs: 'cartesian'})) <= 10
+         |RETURN p.location as point
+        """.stripMargin
+    // When
+    val result = executeWith(distanceConfig, query)
+
+    // Then
+    val plan = result.executionPlanDescription()
+    plan should useOperatorWithText("Projection", "point")
+    plan should useOperatorWithText("Filter", "distance")
+    plan should useOperatorWithText("NodeByLabelScan", ":Place")
+    result.toList.toSet should equal(Set.empty)
+  }
+
+  test("no error for distance with no point") {
+    // Given
+    graph.createIndex("Place", "location")
+    graph.execute("CREATE (p:Place) SET p.location = point({y: 0, x: 0, crs: 'cartesian'})")
+    Range(11, 100).foreach(i => graph.execute(s"CREATE (p:Place) SET p.location = point({y: $i, x: $i, crs: 'cartesian'})"))
+
+    val query =
+      """CYPHER MATCH (p:Place)
+         |WHERE distance(p.location, $poi) <= 10
+         |RETURN p.location as point
+        """.stripMargin
+    // When
+    val result = executeWith(distanceConfig, query, params = Map("poi" -> 5))
+
+    // Then
+    result.toList shouldBe empty
+
+    // And given
+    graph.execute(s"DROP INDEX ON :Place(location)")
+    // when
+    val resultNoIndex = executeWith(distanceConfig, query,  params = Map("poi" -> 5))
+
+    // Then
+    resultNoIndex.toList shouldBe empty
+  }
+}
